WIP Additional Wallet Endpoints *finished tests *changed prefix for the previous wallet endpoints
